Count_Zero_Subarrays.cpp modified and added

+++ C-Plus-Plus/other/Count_Zero_Subarrays.cpp
@@ -1,12 +1,12 @@
 #include <bits/stdc++.h>
 using namespace std;
-typedef long long int ll; 
+typedef long long int ll;
 
 /*
   1. Brute force : TC => O(n^2)
      => Looped over all the subarrays and found out the zero sum subarrays
 */
-ll Count_subarrays_brute(int arr[], int n)
+ll Count_subarrays_brute(ll arr[], ll n)
 {
     ll cnt = 0;
     for (int i = 0; i < n; i++)
@@ -25,7 +25,7 @@ ll Count_subarrays_brute(int arr[], int n)
   2. Optimised solution : TC => O(n)
     => Used Hashmap to keep track of repeated sums which indicates the presence of a zero sum subarray
 */
-ll Count_Subarrays_Optimised(int arr[], int n)
+ll Count_Subarrays_Optimised(ll arr[], ll n)
 {
     ll cnt = 0;
     unordered_map<ll, ll> m;
